+++ src/components/checkout/RemoveProduct.js
